Use public API to create custom BIO in tls-boringssl.c
This avoids reaching into private fields, so that this continues to
build when BIO and BIO_METHOD become opaque. In doing so, I've removed
the puts callback because it is a no-op by multiple counts. (BoringSSL
implements BIO_puts automatically, and the TLS library doesn't use
BIO_puts anyway.)

Likewise, I've removed the new and free functions because they don't do
anything.

Test: treehugger
Change-Id: Ic6c3e988f926d8ca9c466e20c4b76b097e55560b